+++ source/javascripts/components/unified-editor/SwitchBranchDialog/SwitchBranchDialog.tsx
@@ -4,7 +4,7 @@ const SwitchBranchDialog = (props: Omit<DialogProps, 'title'>) => {
   const { isOpen, onClose } = props;
 
   return (
-    <Dialog title="Switch branch" isOpen={isOpen} onClose={onClose}>
+    <Dialog title="Switch branch" isOpen={isOpen} onClose={onClose} as="form">
       <DialogBody>
         <Text>Load configuration from selected branch.</Text>
         <Select label="Branch" placeholder="Select branch" isRequired mt="24" />
@@ -13,7 +13,7 @@ const SwitchBranchDialog = (props: Omit<DialogProps, 'title'>) => {
         <Button onClick={onClose} variant="secondary">
           Cancel
         </Button>
-        <Button>Switch</Button>
+        <Button type="submit">Switch</Button>
       </DialogFooter>
     </Dialog>
   );
